# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API WitelonBank",
        "description": "Dokumentacja API dla systemu bankowego WitelonBank. Pamiętaj o autoryzacji (przycisk 'Authorize' na górze) dla chronionych endpointów.",
        "contact": {
            "email": "support@witelonbank.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Główny serwer API"
        }
    ],
    "paths": {
...
        "/api/zapisani-odbiorcy": {
            "get": {
                "tags": [
                    "Zapisani Odbiorcy"
                ],
                "summary": "Pobiera listę zapisanych odbiorców zalogowanego użytkownika",
                "operationId": "b8f288f1de3bb54be2eb297655521e50",
                "responses": {
                    "200": {
                        "description": "Lista odbiorców",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ZapisanyOdbiorca"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nieautoryzowany"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Zapisani Odbiorcy"
                ],
                "summary": "Dodaje nowego odbiorcę do listy zapisanych",
                "operationId": "41d4958efaf7b395212277eca62cb94f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nazwa_odbiorcy_zdefiniowana",
                                    "nr_konta_odbiorcy"
                                ],
                                "properties": {
                                    "nazwa_odbiorcy_zdefiniowana": {
                                        "type": "string",
                                        "example": "Mieszkanie"
                                    },
                                    "nr_konta_odbiorcy": {
                                        "type": "string",
                                        "example": "PL12345678901234567890123456"
                                    },
                                    "rzeczywista_nazwa_odbiorcy": {
                                        "type": "string",
                                        "example": "Jan Kowalski Wynajem",
                                        "nullable": true
                                    },
                                    "adres_odbiorcy_linia1": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "adres_odbiorcy_linia2": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Odbiorca dodany",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ZapisanyOdbiorca"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/zapisani-odbiorcy/{zapisany_odbiorca}": {
            "get": {
                "tags": [
                    "Zapisani Odbiorcy"
                ],
                "summary": "Pobiera szczegóły zapisanego odbiorcy",
                "operationId": "8dd36896f2e7fce3eff0ded8ae3c4357",
                "parameters": [
                    {
                        "name": "zapisany_odbiorca",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Szczegóły odbiorcy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ZapisanyOdbiorca"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Brak uprawnień"
                    },
                    "404": {
                        "description": "Nie znaleziono"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Zapisani Odbiorcy"
                ],
                "summary": "Aktualizuje zapisanego odbiorcę",
                "operationId": "9256c8a7d4917a9a0afc1e9e8995abc3",
                "parameters": [
                    {
                        "name": "zapisany_odbiorca",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nazwa_odbiorcy_zdefiniowana": {
                                        "type": "string"
                                    },
                                    "nr_konta_odbiorcy": {
                                        "type": "string"
                                    },
                                    "rzeczywista_nazwa_odbiorcy": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "adres_odbiorcy_linia1": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "adres_odbiorcy_linia2": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Odbiorca zaktualizowany",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ZapisanyOdbiorca"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Brak uprawnień"
                    },
                    "404": {
                        "description": "Nie znaleziono"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Zapisani Odbiorcy"
                ],
                "summary": "Usuwa zapisanego odbiorcę",
                "operationId": "be02d8307c985910c0ef836cdb49ef6e",
                "parameters": [
                    {
                        "name": "zapisany_odbiorca",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Odbiorca usunięty"
                    },
                    "403": {
                        "description": "Brak uprawnień"
                    },
                    "404": {
                        "description": "Nie znaleziono"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ZapisanyOdbiorca": {
                "title": "Zapisany Odbiorca",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "id_uzytkownika": {
                        "type": "integer"
                    },
                    "nazwa_odbiorcy_zdefiniowana": {
                        "description": "Nazwa odbiorcy nadana przez użytkownika",
                        "type": "string"
                    },
                    "nr_konta_odbiorcy": {
                        "description": "Numer konta odbiorcy",
                        "type": "string"
                    },
                    "rzeczywista_nazwa_odbiorcy": {
                        "description": "Rzeczywista nazwa odbiorcy (jeśli znana)",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Wprowadź token JWT z prefiksem Bearer (np. 'Bearer eyJhbGciOiJIUzI1NiI...). Token uzyskasz po zalogowaniu.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}